- colibri-console: fix compilaiton and loading of default sequence - generic: fix compilation of OpenNI2 FILE reading but not live acquisition yet - Fix compilation

+++ trunk/main/tools/src/timehistogramwidget.cpp
@@ -94,7 +94,7 @@ void TimeHistogramWidget::displayHisto(t_time_histogram time_histo)
 			swReleaseImage(&mHistoImg);
 			return;
 		}
-		if(time_histo.value_max < time_histo.overflow_count) {
+		if(time_histo.value_max < (unsigned)time_histo.overflow_count) {
 			time_histo.value_max = time_histo.overflow_count ;
 		}
